feat(data): migrate icd10.sdf reference tables to SQLite (Phase A, T012b)
兩階段遷移(net8 無法讀 SQL CE):
- Stage 1 scripts/export_sdf.ps1(32-bit PS + System.Data.SqlServerCe)匯出 5 張
  核心 RDDT 表為 TSV(NULL=\N、UTF-8),共 822,980 列。
- Stage 2 tools/Drg.Migrate 載入 TSV → icd10.sqlite(數值欄以 SQLite 親和性轉換)。

RulesetRepository 修正表名 RDDT_ECC→RDDT_ECC_V / RDDT_ECC_GROUP→RDDT_ECC_GROUP_V;
MdcDrgWgt/MdcDrgXicd 數值欄改 long/double 對齊 SQLite(Dapper positional-record)。

新增 Drg.Data.Tests 整合測試:對真實 icd10.sqlite 驗證 ruleset 載入(筆數/NULL 對映)
與 MdcCheck 對真實主診斷碼的 MDC 指派——第一個真實 parity 訊號。衍生資料
(migration/、icd10.sqlite)gitignore。全套 72 passed / 1 skip。

+++ src/Drg.Core/Engine/TreeSelector.cs
@@ -12,8 +12,8 @@ public static class TreeSelector
         Remap(drg);
         var set = drg.Where(d => d.Length > 0).ToHashSet();
 
-        var bestNo = 200;
-        var bestWgt = 0f;
+        var bestNo = 200L;
+        var bestWgt = 0d;
         var result = "";
         foreach (var r in rs.MdcDrgWgt)
         {
@@ -27,7 +27,7 @@ public static class TreeSelector
                 if (treeNo < bestNo) { bestNo = treeNo; bestWgt = wgt; result = treeDrg; }
                 continue;
             }
-            if (wgt == 0f) wgt = (float)ctx.MedAmt / r.AvgExp;   // legacy:零權重以平均費用回推
+            if (wgt == 0d) wgt = (double)ctx.MedAmt / r.AvgExp;   // legacy:零權重以平均費用回推
             if (wgt > bestWgt || (wgt == bestWgt && treeNo < bestNo))
             {
                 bestNo = treeNo;
